requirements.txt
langchain
langchain-core
langchain-openai
langchain-groq
python-dotenv